+++ apps/server/src/api/user/dto/generate-auth-passkey.dto.ts
@@ -1,6 +1,6 @@
 import { IsEmail } from "class-validator"
 
-export class GenerateIdentityPasskeyDto {
+export class GenerateAuthPasskeyDto {
   @IsEmail()
   readonly email: string
 }